requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.0
attrs==21.4.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
clarifai-grpc==7.11.0
colorama==0.4.4
dnspython==2.2.0
emoji==1.6.3
ephem==4.1.3
flake8==4.0.1
googleapis-common-protos==1.54.0
grpcio==1.43.0
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
mongomock==4.0.0
packaging==21.3
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pymongo==4.0.1
pyparsing==3.0.8
PySocks==1.7.1
pytest==7.1.2
pytest-mongodb==2.2.0
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
sentinels==1.0.0
six==1.16.0
tomli==2.0.1
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
